requirements.txt
# ==============================================
# QE Orchestrator - Python Dependencies
# ==============================================

# Core Framework
langgraph==0.2.16
langchain-core==0.2.38
langchain==0.2.14

# AI/ML Integrations
google-generativeai==0.8.0

# External Service Clients
jira==3.5.2
pygithub==2.1.1

# Web Framework (FastAPI + Uvicorn)
fastapi==0.109.0
uvicorn[standard]==0.27.0
sse-starlette==2.0.0
python-multipart==0.0.6

# Configuration & Environment
python-dotenv==1.0.0

# Data Validation & Serialization
pydantic>=2.7.4
pydantic-settings==2.1.0

# HTTP & Networking
requests==2.31.0
urllib3==1.26.18

# Utilities
python-dateutil==2.8.2
pytz==2024.1

# Testing (Development)
pytest==7.4.3
pytest-mock==3.12.0
pytest-cov==4.1.0
pytest-asyncio==0.23.3

# Code Quality (Development)
black==24.1.1
flake8==7.0.0
pylint==3.0.3
mypy==1.8.0

# Security
cryptography==42.0.0

# Document Generation
reportlab==4.0.7
openpyxl==3.1.2